+++ lib/eo_color.dart
@@ -1,7 +1,2 @@
-library eo_color;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/palette.dart';
+export 'src/palette_rgb.dart';