+++ lib/src/palette.dart
@@ -1,0 +1,7 @@
+import 'dart:ui';
+
+/// Color palette from which a color is picked.
+abstract class Palette {
+  /// The picked color.
+  Color get color;
+}